test(admin): split disk-usage OPSEC tests out of coverage_api_routers
The pytest fastlane plans the changed test module; editing
tests/test_coverage_api_routers.py collected the whole file, including
TestPDFPool which patches pymupdf — absent from slim fastlane deps —
failing with ModuleNotFoundError. Move TestAdminDiskUsage into a
dedicated tests/test_admin_disk_usage_opsec.py with minimal fixtures.
TestPDFPool is unchanged.

X-Agent: kimi/infra-7081-fastlane
